requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP Requests
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Date/Time
python-dateutil==2.8.2

# Utilities
tenacity==8.2.3

# LLM Narratives (Phase 2)
anthropic==0.40.0

# Testing (optional for now)
pytest==7.4.4
pytest-asyncio==0.23.3
slowapi==0.1.9
